Extract test data to a separate file
The test data (here in the main function) are extracted from the code
to another file, so it is easier to understand what file serves what
purpose.
